feat(routing): promote live data page to /dashboard, move event detail to /dashboard/events/data/[eventId]
Closes #67

- Replace mock KPI dashboard at /dashboard with the live Convex attendance analytics page
- Move event attendance detail from /dashboard/data/[eventId] to /dashboard/events/data/[eventId]
- Delete entire app/dashboard/data/ directory (no redirect)
- Update back button on detail page: href /dashboard/data → /dashboard, label "Back to Data" → "Back to Dashboard"
- Remove Data nav item and BarChart3 import from layout.tsx
